+++ src/bhad/model.py
@@ -1,10 +1,10 @@
-from sklearn.base import BaseEstimator, OutlierMixin
-from sklearn.preprocessing import OneHotEncoder
-from typing import (List, Optional)
-import numpy as np
-import pandas as pd
+from typing import (List, Optional, Union)
 from copy import deepcopy
 import warnings
+import pandas as pd
+import numpy as np
+from sklearn.base import BaseEstimator, OutlierMixin
+#from sklearn.preprocessing import OneHotEncoder
 import bhad.utils as utils
 
 class BHAD(BaseEstimator, OutlierMixin):
@@ -57,22 +57,29 @@ class BHAD(BaseEstimator, OutlierMixin):
     [1] Vosseler, A. (2021): BHAD: Fast unsupervised anomaly detection using Bayesian histograms, Technical Report
     """
 
-    def __init__(self, contamination : float = 0.01, alpha : float = 1/2, exclude_col : Optional[List[str]] = [], append_score : bool = False, verbose : bool = True):
+    def __init__(self, contamination : float = 0.01, alpha : float = 1/2, exclude_col : Optional[List[str]] = [], 
+                    numeric_features : Optional[List[str]] = [], cat_features : Optional[List[str]] = [],
+                    append_score : bool = False, verbose : bool = True):
         
+        super(BHAD, self).__init__()
         self.contamination = contamination              # outlier proportion in the dataset
         self.alpha = alpha                              # uniform Dirichlet prior concentration parameter used for each feature
         self.verbose = verbose
         self.append_score = append_score
+        self.numeric_features = numeric_features
+        self.cat_features = cat_features 
         self.exclude_col = exclude_col               # list with column names in X of columns to exclude for computation of the score
-        if self.verbose : print("\n-- Bayesian Histogram-based Anomaly Detector (BHAD) --\n")
-        super(BHAD, self).__init__()
+        if self.verbose : 
+            print("\n-- Bayesian Histogram-based Anomaly Detector (BHAD) --\n")
+        
 
     def __del__(self):
         class_name = self.__class__.__name__
 
     def __repr__(self):
-        return f"BHAD(contamination = {self.contamination}, alpha = {self.alpha})"
+        return f"BHAD(contamination = {self.contamination}, alpha = {self.alpha}, exclude_col = {self.exclude_col}, numeric_features = {self.numeric_features}, cat_features = {self.cat_features}, append_score = {self.append_score}, verbose = {self.verbose})"
     
+
     def _fast_bhad(self, X : pd.DataFrame)-> pd.DataFrame:
       """
       Input:
@@ -84,13 +91,13 @@ class BHAD(BaseEstimator, OutlierMixin):
       assert isinstance(X, pd.DataFrame), 'X must be of type pd.DataFrame'
       selected_col = X.columns[~X.columns.isin(self.exclude_col)] 
       if len(self.exclude_col)>0:
-         print("Features",self.exclude_col, 'excluded.')  
+            print("Features",self.exclude_col, 'excluded.')  
         
-      df = deepcopy(X[selected_col]) 
+      df = deepcopy(X[selected_col].astype(object)) 
       self.df_shape = df.shape  
-      self.columns = df.select_dtypes(include='object').columns.tolist()  # use only categorical (including discretized numerical)
+      self.columns = df.select_dtypes(include=['object', 'category']).columns.tolist()  # use only categorical (including discretized numerical)
       if len(self.columns)!= self.df_shape[1] : 
-        warnings.warn('Not all features in X are categorical!!')
+            warnings.warn('Not all features in X are categorical!!')
       self.df = df
       unique_categories_ = [df[var].unique().tolist() + ['infrequent'] for var in df.columns]
       
@@ -100,7 +107,8 @@ class BHAD(BaseEstimator, OutlierMixin):
       self.enc = utils.onehot_encoder(prefix_sep='__', verbose=self.verbose)   # more flexible but much slower
       self.df_one = self.enc.fit_transform(df).toarray()   
       assert all(np.sum(self.df_one, axis=1) == df.shape[1]), 'Row sums must be equal to number of features!!'
-      if self.verbose : print("Matrix dimension after one-hot encoding:", self.df_one.shape)  
+      if self.verbose : 
+        print("Matrix dimension after one-hot encoding:", self.df_one.shape)  
       self.columns_onehot_ = self.enc.get_feature_names_out()
       
       # Prior parameters and sufficient statistics:
@@ -121,18 +129,18 @@ class BHAD(BaseEstimator, OutlierMixin):
       # Assign each obs. the overall category count
       #-------------------------------------------------------------------
       self.f_mat = self.df_one * np.array(a)              # keep only nonzero matrix entries
-      f_mat_bayes = self.df_one * np.array(a_bayes)
+      self.f_mat_bayes = self.df_one * np.array(a_bayes)
 
       # Calculate outlier score for each row (observation), 
       # see equation (5) in [1]
       #-----------------------------------------------------
-      out = pd.Series(np.apply_along_axis(np.sum, 1, f_mat_bayes), index=df.index)    
+      out = pd.Series(np.apply_along_axis(np.sum, 1, self.f_mat_bayes), index=df.index)    
       if self.append_score:  
          out = pd.concat([df, pd.DataFrame(out, columns = ['outlier_score'])], axis=1)
       return out    
     
-    @utils.timer
-    def fit(self, X : pd.DataFrame, y=None)-> 'BHAD':
+
+    def fit(self, X : pd.DataFrame, y : Union[np.array, pd.Series] = None)-> 'BHAD':
         """
         Apply the BHAD and calculate the outlier threshold value.
 
@@ -147,23 +155,28 @@ class BHAD(BaseEstimator, OutlierMixin):
         Returns
         -------
         self : BHAD object
-        """
+        """                
         # Fit model: 
         #------------
-        if self.verbose : print("Fit BHAD on discretized data.")
+        if self.verbose : 
+            print("Fit BHAD on discretized data.")
+            print(f"Input shape: {X.shape}")
+        
         self.scores = self._fast_bhad(X)
+
         if self.append_score:  
             self.threshold_ = np.nanpercentile(self.scores['outlier_score'].tolist(), q=100*self.contamination)
         else: 
             self.threshold_ = np.nanpercentile(self.scores.tolist(), q=100*self.contamination)
-        if self.verbose : print("Finished training.")
+        if self.verbose : 
+            print("Finished training.")
         
         # Tag as fitted for sklearn compatibility: 
         # https://scikit-learn.org/stable/developers/develop.html#estimated-attributes
         self.X_ = X
         self.xindex_fitted_, self.df_, self.scores_, self.freq_ = self.X_.index, self.df, self.scores, self.freq          
-        self.enc_, self.df_one_, self.f_mat_ = self.enc, self.df_one, self.f_mat
-        del self.df_one, self.f_mat, self.freq, self.df
+        self.enc_, self.df_one_, self.f_mat_, self.f_mat_bayes_ = self.enc, self.df_one, self.f_mat, self.f_mat_bayes
+        self.numeric_features_, self.cat_features_ = self.numeric_features, self.cat_features
         return self
 
     
@@ -184,9 +197,10 @@ class BHAD(BaseEstimator, OutlierMixin):
             The outlier score of the input samples centered arount threshold 
             value.
         """
-        if self.verbose : print("\nScore input data.\nDiscretize continous features.")
-        df = deepcopy(X)
-        if self.verbose : print("Apply fitted one-hot encoder.")        
+        if self.verbose : 
+            print("\nScore input data.")
+            print("Apply fitted one-hot encoder.")        
+        df = deepcopy(X)    
         self.df_one = self.enc_.transform(df).toarray()     # apply fitted one-hot encoder to categorical -> sparse dummy matrix
         assert all(np.sum(self.df_one, axis=1) == df.shape[1]), 'Row sums must be equal to number of features!!'
         
@@ -227,7 +241,8 @@ class BHAD(BaseEstimator, OutlierMixin):
         """
         # Center scores; divide into outlier and inlier (-/+)
         if hasattr(self, 'X_') and X.equals(self.X_): 
-            if self.verbose : print("Score input data.")
+            if self.verbose : 
+                print("Score input data.")
             self.anomaly_scores = self.scores_.to_numpy() - self.threshold_
         else:    
             self.anomaly_scores = self.score_samples(X).to_numpy() - self.threshold_